requirements.txt
Flask
Gunicorn
flask-cors
requests
python-dotenv